requirements.txt
Django==1.6.1
Markdown==2.3.1
South==0.8.4
django-extensions==1.2.5
django-markupfield==1.2.0
psycopg2==2.5.1
six==1.4.1
wsgiref==0.1.2
